Adding Hilt for DI support Move MainActivityViewModel.kt to the right package
