+++ app/src/main/java/com/ayogeshwaran/githubclient/GithubClientApp.kt
@@ -1,5 +1,7 @@
 package com.ayogeshwaran.githubclient
 
 import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
 
+@HiltAndroidApp
 class GithubClientApp : Application()
