+++ app/src/main/java/com/ayogeshwaran/githubclient/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.ayogeshwaran.githubclient.closedpr.ClosedPrFragment
-import com.ayogeshwaran.githubclient.closedpr.MainActivityViewModel
 import com.ayogeshwaran.githubclient.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
